Working with some cosmetic fixes Nov21

+++ src/qfi_ADI.cpp
@@ -234,14 +234,14 @@ void qfi_ADI::reset()
 
 void qfi_ADI::updateView()
 {
-    m_scaleX = (float)width()  / (float)m_originalWidth;
-    m_scaleY = (float)height() / (float)m_originalHeight;
+    m_scaleX = static_cast<float>(width())  / static_cast<float>(m_originalWidth);
+    m_scaleY = static_cast<float>(height()) / static_cast<float>(m_originalHeight);
 
     m_itemBack->setRotation( - m_roll );
     m_itemFace->setRotation( - m_roll );
     m_itemRing->setRotation( - m_roll );
 
-    float roll_rad = M_PI * m_roll / 180.0;
+    float roll_rad = static_cast<float>(M_PI) * m_roll / 180.0f;
 
     float delta  = m_originalPixPerDeg * m_pitch;
